+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `AddressContainer.addressStyleRules` prop, starting with `addressStyleRules.requiredIndicator` to fit the form into different design patterns.
+- `AddressContainer.fieldsStyleRules` prop, starting with `fieldsStyleRules.requiredIndicator` to fit the form into different design patterns.
 
 ## [4.10.1] - 2021-10-06